Working class separation with special classes built on top of pydantic models to be able to use custom methods using existing pydantic models, consistent class parameters show() method
