+++ atlast_sc/parameters/instrument_specific_parameters.py
@@ -29,4 +29,4 @@ class InstrumentSpecificParameters:
             attr = getattr(self.__class__, name)
             if isinstance(attr, property):
                 value = getattr(self, name)
-                print(f"  {name}: {value}")
+                print(f"{name}: {value}")
