+++ atlast_sc/parameters/telescope_and_environment_parameters.py
@@ -84,4 +84,4 @@ class TelescopeAndEnvironmentParameters:
             attr = getattr(self.__class__, name)
             if isinstance(attr, property):
                 value = getattr(self, name)
-                print(f"  {name}: {value}")
+                print(f"{name}: {value}")
